refactor(egopulse): migrate turn.rs tests to guards and compaction submodules
