+++ egopulse/src/agent_loop/turn.rs
@@ -879,14 +879,14 @@ pub(crate) fn build_state_with_provider(
 #[cfg(test)]
 mod tests {
     use super::{
-        FakeProvider, FailingProvider, RecordingProvider, build_state, build_state_with_provider,
-        cli_context, test_config_with_compaction,
+        FakeProvider, FailingProvider, RecordingProvider, build_state_with_provider,
+        cli_context,
     };
     use serial_test::serial;
 
     use crate::agent_loop::process_turn;
-    use crate::error::{EgoPulseError, LlmError};
-    use crate::llm::{Message, MessagesResponse, ToolCall};
+    use crate::error::EgoPulseError;
+    use crate::llm::{MessagesResponse, ToolCall};
     use crate::storage::call_blocking;
 
     #[tokio::test]
@@ -980,80 +980,6 @@ mod tests {
 
     #[tokio::test]
     #[serial]
-    async fn empty_reply_guard_retries_once_then_errors() {
-        // LLM returns empty content twice → first triggers guard retry, second errors
-        let dir = tempfile::tempdir().expect("tempdir");
-        let state = build_state_with_provider(
-            dir.path().to_str().expect("utf8").to_string(),
-            Box::new(FakeProvider {
-                responses: std::sync::Mutex::new(vec![
-                    MessagesResponse {
-                        content: String::new(),
-                        tool_calls: Vec::new(),
-                    },
-                    MessagesResponse {
-                        content: String::new(),
-                        tool_calls: Vec::new(),
-                    },
-                ]),
-            }),
-        );
-
-        let error = process_turn(&state, &cli_context("empty-guard"), "hello")
-            .await
-            .expect_err("should fail after retry");
-        assert!(matches!(error, EgoPulseError::Llm(_)));
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn declarative_only_guard_retries_then_returns() {
-        // First: declarative-only reply, Second: actual answer after guard retry
-        let dir = tempfile::tempdir().expect("tempdir");
-        let state = build_state_with_provider(
-            dir.path().to_str().expect("utf8").to_string(),
-            Box::new(FakeProvider {
-                responses: std::sync::Mutex::new(vec![
-                    MessagesResponse {
-                        content: "Sure, I'll help you with that.".to_string(),
-                        tool_calls: Vec::new(),
-                    },
-                    MessagesResponse {
-                        content: "Here is the answer you need.".to_string(),
-                        tool_calls: Vec::new(),
-                    },
-                ]),
-            }),
-        );
-
-        let reply = process_turn(&state, &cli_context("declarative-guard"), "help me")
-            .await
-            .expect("should succeed after retry");
-        assert_eq!(reply, "Here is the answer you need.");
-
-        let chat_id = call_blocking(state.db.clone(), move |db| {
-            db.resolve_or_create_chat_id(
-                "cli",
-                "cli:declarative-guard",
-                Some("declarative-guard"),
-                "cli",
-            )
-        })
-        .await
-        .expect("chat id");
-        let loaded = crate::agent_loop::session::load_messages_for_turn(&state, chat_id)
-            .await
-            .expect("loaded session");
-        assert!(
-            loaded
-                .messages
-                .iter()
-                .all(|message| !message.content.as_text_lossy().contains("[runtime_guard]"))
-        );
-    }
-
-    #[tokio::test]
-    #[serial]
     async fn normal_tool_flow_still_works_after_port() {
         // Regression: existing tool flow with multiple tool calls should still work
         let dir = tempfile::tempdir().expect("tempdir");
@@ -1116,232 +1042,4 @@ mod tests {
             .expect_err("should fail with malformed tool calls");
         assert!(matches!(error, EgoPulseError::Llm(_)));
     }
-
-    #[tokio::test]
-    #[serial]
-    async fn malformed_declarative_tool_reply_retries_then_returns() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let state = build_state_with_provider(
-            dir.path().to_str().expect("utf8").to_string(),
-            Box::new(FakeProvider {
-                responses: std::sync::Mutex::new(vec![
-                    MessagesResponse {
-                        content: "了解しました。実行します。".to_string(),
-                        tool_calls: vec![ToolCall {
-                            id: "call-malformed".to_string(),
-                            name: String::new(),
-                            arguments: serde_json::json!({}),
-                        }],
-                    },
-                    MessagesResponse {
-                        content: "実行結果です。".to_string(),
-                        tool_calls: Vec::new(),
-                    },
-                ]),
-            }),
-        );
-
-        let reply = process_turn(&state, &cli_context("malformed-declarative"), "test")
-            .await
-            .expect("should recover after retry");
-        assert_eq!(reply, "実行結果です。");
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn compaction_summarizes_old_messages_and_persists_summary_context() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let provider = RecordingProvider::new(
-            vec![
-                Ok(MessagesResponse {
-                    content: "summary text".to_string(),
-                    tool_calls: Vec::new(),
-                }),
-                Ok(MessagesResponse {
-                    content: "final answer".to_string(),
-                    tool_calls: Vec::new(),
-                }),
-            ],
-            vec![0, 0],
-        );
-        let config =
-            test_config_with_compaction(dir.path().to_str().expect("utf8").to_string(), 4, 2);
-        let state = build_state(config, Box::new(provider.clone()));
-        let context = cli_context("compaction-success");
-        let chat_id = call_blocking(state.db.clone(), move |db| {
-            db.resolve_or_create_chat_id(
-                "cli",
-                "cli:compaction-success",
-                Some("compaction-success"),
-                "cli",
-            )
-        })
-        .await
-        .expect("chat id");
-        let seeded = vec![
-            Message::text("user", "old-user-1"),
-            Message::text("assistant", "old-assistant-1"),
-            Message::text("user", "old-user-2"),
-            Message::text("assistant", "old-assistant-2"),
-        ];
-        let seeded_json = serde_json::to_string(&seeded).expect("seeded json");
-        call_blocking(state.db.clone(), move |db| {
-            db.save_session(chat_id, &seeded_json)
-        })
-        .await
-        .expect("save session");
-
-        let reply = process_turn(&state, &context, "fresh question")
-            .await
-            .expect("process turn");
-        assert_eq!(reply, "final answer");
-
-        let seen_systems = provider.seen_systems();
-        assert_eq!(seen_systems.len(), 2);
-        assert_eq!(seen_systems[0], "You are a helpful summarizer.");
-
-        let seen_messages = provider.seen_messages();
-        assert_eq!(seen_messages.len(), 2);
-        assert_eq!(
-            seen_messages[1][0].content.as_text_lossy(),
-            "[Conversation Summary]\nsummary text"
-        );
-        assert_eq!(seen_messages[1][1].role, "assistant");
-        assert_eq!(
-            seen_messages[1][1].content.as_text_lossy(),
-            "old-assistant-2"
-        );
-        assert_eq!(
-            seen_messages[1]
-                .last()
-                .expect("final request")
-                .content
-                .as_text_lossy(),
-            "fresh question"
-        );
-
-        let loaded = crate::agent_loop::session::load_messages_for_turn(&state, chat_id)
-            .await
-            .expect("loaded session");
-        assert_eq!(
-            loaded.messages[0].content.as_text_lossy(),
-            "[Conversation Summary]\nsummary text"
-        );
-        assert_eq!(
-            loaded
-                .messages
-                .last()
-                .expect("session last")
-                .content
-                .as_text_lossy(),
-            "final answer"
-        );
-
-        let archive_dir = dir
-            .path()
-            .join("groups")
-            .join("cli")
-            .join(chat_id.to_string())
-            .join("conversations");
-        let archives = std::fs::read_dir(&archive_dir)
-            .expect("archive dir")
-            .collect::<Result<Vec<_>, _>>()
-            .expect("archive entries");
-        assert_eq!(archives.len(), 1);
-        let archive_body = std::fs::read_to_string(archives[0].path()).expect("archive body");
-        assert!(archive_body.contains("old-user-1"));
-        assert!(archive_body.contains("fresh question"));
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn compaction_falls_back_to_recent_messages_when_summary_fails() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let provider = RecordingProvider::new(
-            vec![
-                Err(LlmError::InvalidResponse("summary failed".to_string())),
-                Ok(MessagesResponse {
-                    content: "final answer".to_string(),
-                    tool_calls: Vec::new(),
-                }),
-            ],
-            vec![0, 0],
-        );
-        let config =
-            test_config_with_compaction(dir.path().to_str().expect("utf8").to_string(), 4, 2);
-        let state = build_state(config, Box::new(provider.clone()));
-        let context = cli_context("compaction-fallback");
-        let chat_id = call_blocking(state.db.clone(), move |db| {
-            db.resolve_or_create_chat_id(
-                "cli",
-                "cli:compaction-fallback",
-                Some("compaction-fallback"),
-                "cli",
-            )
-        })
-        .await
-        .expect("chat id");
-        let seeded = vec![
-            Message::text("user", "old-user-1"),
-            Message::text("assistant", "old-assistant-1"),
-            Message::text("user", "old-user-2"),
-            Message::text("assistant", "old-assistant-2"),
-        ];
-        let seeded_json = serde_json::to_string(&seeded).expect("seeded json");
-        call_blocking(state.db.clone(), move |db| {
-            db.save_session(chat_id, &seeded_json)
-        })
-        .await
-        .expect("save session");
-
-        let reply = process_turn(&state, &context, "fresh question")
-            .await
-            .expect("process turn");
-        assert_eq!(reply, "final answer");
-
-        let seen_messages = provider.seen_messages();
-        assert_eq!(seen_messages.len(), 2);
-        assert!(seen_messages[1].iter().all(|message| {
-            !message
-                .content
-                .as_text_lossy()
-                .contains("[Conversation Summary]")
-        }));
-        assert_eq!(
-            seen_messages[1][0].content.as_text_lossy(),
-            "old-assistant-2"
-        );
-        assert_eq!(
-            seen_messages[1]
-                .last()
-                .expect("final request")
-                .content
-                .as_text_lossy(),
-            "fresh question"
-        );
-
-        let loaded = crate::agent_loop::session::load_messages_for_turn(&state, chat_id)
-            .await
-            .expect("loaded session");
-        assert!(loaded.messages.iter().all(|message| {
-            !message
-                .content
-                .as_text_lossy()
-                .contains("[Conversation Summary]")
-        }));
-        assert_eq!(
-            loaded.messages[0].content.as_text_lossy(),
-            "old-assistant-2"
-        );
-        assert_eq!(
-            loaded
-                .messages
-                .last()
-                .expect("session last")
-                .content
-                .as_text_lossy(),
-            "final answer"
-        );
-    }
-
 }
